# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Pact Provider Demo API
  description: Provider service for AI-powered contract testing demonstration
  version: 1.0.0
  contact:
    name: Contract Testing Team

servers:
  - url: http://localhost:3001
    description: Local development

tags:
  - name: Items
    description: Item management operations
  - name: Categories
    description: Category management operations
  - name: Users
    description: User management operations
  - name: System
    description: System health and status operations

paths:
  /items:
    get:
      summary: List all items
      description: Retrieve a list of all items with optional filtering
      operationId: getItems
      tags:
        - Items
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: inStock
          in: query
          schema:
            type: boolean
          description: Filter by stock status
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  total:
                    type: integer

    post:
      summary: Create a new item
      description: Create a new item in the inventory
      operationId: createItem
      tags:
        - Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /items/search:
    get:
      summary: Search items
      description: Search for items by query string
      operationId: searchItems
      tags:
        - Items
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  query:
                    type: string
                  count:
                    type: integer
        '400':
          description: Missing query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /items/{id}:
    get:
      summary: Get item by ID
      description: Retrieve a single item by its ID
      operationId: getItemById
      tags:
        - Items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Replace item
      description: Replace an entire item with new data
      operationId: replaceItem
      tags:
        - Items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
      responses:
        '200':
          description: Item replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found

    patch:
      summary: Update item
      description: Partially update an item's properties
      operationId: updateItem
      tags:
        - Items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found

    delete:
      summary: Delete item
      description: Delete an item from the inventory
      operationId: deleteItem
      tags:
        - Items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Item deleted
        '404':
          description: Item not found

  /categories:
    get:
      summary: List all categories
      description: Retrieve a list of all categories
      operationId: getCategories
      tags:
        - Categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
                  total:
                    type: integer

    post:
      summary: Create category
      description: Create a new category
      operationId: createCategory
      tags:
        - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                slug:
                  type: string
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'

  /categories/{id}:
    get:
      summary: Get category by ID
      description: Retrieve a single category by its ID
      operationId: getCategoryById
      tags:
        - Categories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found

  /categories/{id}/items:
    get:
      summary: Get items in category
      description: Retrieve all items belonging to a specific category
      operationId: getCategoryItems
      tags:
        - Categories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Items in category
          content:
            application/json:
              schema:
                type: object
                properties:
                  category:
                    type: string
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  count:
                    type: integer
        '404':
          description: Category not found

  /users/{id}:
    get:
      summary: Get user by ID
      description: Retrieve public user information by ID
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details (without email)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '404':
          description: User not found

  /users/{id}/profile:
    get:
      summary: Get user profile
      description: Retrieve full user profile including email
      operationId: getUserProfile
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User profile (with email)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

  /health:
    get:
      summary: Health check
      description: Check if the service is running and healthy
      operationId: healthCheck
      tags:
        - System
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time

components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        category:
          type: string
        inStock:
          type: boolean

    ItemInput:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
        price:
          type: number
        category:
          type: string
        inStock:
          type: boolean

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        itemCount:
          type: integer

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string

    UserPublic:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string
        id:
          type: integer